+++ Engine/Source/Rendering/Native/Pipelines/GraphicsPipelines/SkyGraphicsPipeline.cpp
@@ -119,7 +119,7 @@ void SkyGraphicsPipeline::Execute() NOEXCEPT
 
 		for (uint64 i{ 0 }; i < number_of_light_components; ++i, ++component)
 		{
-			if (component->_LightType == static_cast<uint32>(LightType::DIRECTIONAL))
+			if (component->_LightType == LightType::DIRECTIONAL)
 			{
 				data._SkyLightDirection = -component->_Direction;
 				data._SkyLightRadiance = component->_Color * component->_Intensity;
