+++ Engine/Source/Rendering/Native/RenderPasses/ShadowsRenderPass.cpp
@@ -203,7 +203,7 @@ void ShadowsRenderPass::Execute() NOEXCEPT
 
 		for (uint64 i{ 0 }; i < number_of_light_components; ++i, ++component)
 		{
-			if (component->_LightType == static_cast<uint32>(LightType::DIRECTIONAL))
+			if (component->_LightType == LightType::DIRECTIONAL)
 			{
 				for (uint8 i{ 0 }; i < 4; ++i)
 				{
